Move JobURL into pjutil so controllers can reuse it
