+++ prow/pjutil/BUILD.bazel
@@ -17,6 +17,7 @@ go_library(
     importpath = "k8s.io/test-infra/prow/pjutil",
     deps = [
         "//prow/config:go_default_library",
+        "//prow/gcsupload:go_default_library",
         "//prow/github:go_default_library",
         "//prow/kube:go_default_library",
         "//prow/pod-utils/decorate:go_default_library",
@@ -51,6 +52,7 @@ go_test(
     deps = [
         "//prow/config:go_default_library",
         "//prow/kube:go_default_library",
+        "//vendor/github.com/sirupsen/logrus:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/equality:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/diff:go_default_library",
