+++ prow/pjutil/pjutil.go
@@ -18,14 +18,20 @@ limitations under the License.
 package pjutil
 
 import (
+	"bytes"
+	"net/url"
+	"path"
+
 	"github.com/satori/go.uuid"
 	"github.com/sirupsen/logrus"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"k8s.io/test-infra/prow/config"
+	"k8s.io/test-infra/prow/gcsupload"
 	"k8s.io/test-infra/prow/github"
 	"k8s.io/test-infra/prow/kube"
 	"k8s.io/test-infra/prow/pod-utils/decorate"
+	"k8s.io/test-infra/prow/pod-utils/downwardapi"
 )
 
 // NewProwJobWithAnnotation initializes a ProwJob out of a ProwJobSpec with annotations.
@@ -234,4 +240,26 @@ func ProwJobFields(pj *kube.ProwJob) logrus.Fields {
 		fields[github.OrgLogField] = pj.Spec.Refs.Org
 	}
 	return fields
+}
+
+// JobURL returns the expected URL for ProwJobStatus.
+//
+// TODO(fejta): consider moving default JobURLTemplate and JobURLPrefix out of plank
+func JobURL(plank config.Plank, pj kube.ProwJob, log *logrus.Entry) string {
+	if pj.Spec.DecorationConfig != nil && plank.JobURLPrefix != "" {
+		spec := downwardapi.NewJobSpec(pj.Spec, pj.Status.BuildID, pj.Name)
+		gcsConfig := pj.Spec.DecorationConfig.GCSConfiguration
+		_, gcsPath, _ := gcsupload.PathsForJob(gcsConfig, &spec, "")
+
+		prefix, _ := url.Parse(plank.JobURLPrefix)
+		prefix.Path = path.Join(prefix.Path, gcsConfig.Bucket, gcsPath)
+		return prefix.String()
+	}
+	var b bytes.Buffer
+	if err := plank.JobURLTemplate.Execute(&b, &pj); err != nil {
+		log.WithFields(ProwJobFields(&pj)).Errorf("error executing URL template: %v", err)
+	} else {
+		return b.String()
+	}
+	return ""
 }
