+++ prow/plank/BUILD.bazel
@@ -27,12 +27,10 @@ go_library(
     importpath = "k8s.io/test-infra/prow/plank",
     deps = [
         "//prow/config:go_default_library",
-        "//prow/gcsupload:go_default_library",
         "//prow/github:go_default_library",
         "//prow/kube:go_default_library",
         "//prow/pjutil:go_default_library",
         "//prow/pod-utils/decorate:go_default_library",
-        "//prow/pod-utils/downwardapi:go_default_library",
         "//prow/report:go_default_library",
         "//vendor/github.com/sirupsen/logrus:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
