+++ prow/plank/controller.go
@@ -17,23 +17,19 @@ limitations under the License.
 package plank
 
 import (
-	"bytes"
 	"fmt"
-	"net/url"
-	"path"
 	"strings"
 	"sync"
 	"time"
 
 	"github.com/sirupsen/logrus"
 	"k8s.io/api/core/v1"
+
 	"k8s.io/test-infra/prow/config"
-	"k8s.io/test-infra/prow/gcsupload"
 	"k8s.io/test-infra/prow/github"
 	"k8s.io/test-infra/prow/kube"
 	"k8s.io/test-infra/prow/pjutil"
 	"k8s.io/test-infra/prow/pod-utils/decorate"
-	"k8s.io/test-infra/prow/pod-utils/downwardapi"
 	reportlib "k8s.io/test-infra/prow/report"
 )
 
@@ -417,7 +413,7 @@ func (c *Controller) syncPendingJob(pj kube.ProwJob, pm map[string]kube.Pod, rep
 		}
 	}
 
-	pj.Status.URL = jobURL(c.ca.Config().Plank, pj, c.log)
+	pj.Status.URL = pjutil.JobURL(c.ca.Config().Plank, pj, c.log)
 
 	reports <- pj
 
@@ -469,7 +465,7 @@ func (c *Controller) syncTriggeredJob(pj kube.ProwJob, pm map[string]kube.Pod, r
 		pj.Status.State = kube.PendingState
 		pj.Status.PodName = pn
 		pj.Status.Description = "Job triggered."
-		pj.Status.URL = jobURL(c.ca.Config().Plank, pj, c.log)
+		pj.Status.URL = pjutil.JobURL(c.ca.Config().Plank, pj, c.log)
 	}
 	reports <- pj
 	if prevState != pj.Status.State {
@@ -554,23 +550,4 @@ func (c *Controller) RunAfterSuccessCanRun(parent, child *kube.ProwJob, ca confi
 		changes = append(changes, change.Filename)
 	}
 	return ps.RunsAgainstChanges(changes)
-}
-
-func jobURL(plank config.Plank, pj kube.ProwJob, log *logrus.Entry) string {
-	if pj.Spec.DecorationConfig != nil && plank.JobURLPrefix != "" {
-		spec := downwardapi.NewJobSpec(pj.Spec, pj.Status.BuildID, pj.Name)
-		gcsConfig := pj.Spec.DecorationConfig.GCSConfiguration
-		_, gcsPath, _ := gcsupload.PathsForJob(gcsConfig, &spec, "")
-
-		prefix, _ := url.Parse(plank.JobURLPrefix)
-		prefix.Path = path.Join(prefix.Path, gcsConfig.Bucket, gcsPath)
-		return prefix.String()
-	}
-	var b bytes.Buffer
-	if err := plank.JobURLTemplate.Execute(&b, &pj); err != nil {
-		log.WithFields(pjutil.ProwJobFields(&pj)).Errorf("error executing URL template: %v", err)
-	} else {
-		return b.String()
-	}
-	return ""
 }
